% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granby.R
\docType{data}
\name{granby}
\alias{granby}
\title{Granby BBS route.}
\format{
A MATSS dataset.
\describe{
  \item{abundance}{tibble of species abundances (columns) per year (rows)}
  \item{covariates}{tibble of covariate variables (columns) per year (rows)}
  \item{metadata}{metadata list}
}
}
\usage{
granby
}
\description{
As loaded from MATSS (`get_bbs_route_region_data`). See MATSS for full documentation.
}
\keyword{datasets}
