% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{list_species}
\alias{list_species}
\title{List species}
\usage{
list_species(a_dataset)
}
\arguments{
\item{a_dataset}{matss dataset}
}
\value{
list of species names
}
\description{
List species
}
