% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_methods.R
\name{make_ibd}
\alias{make_ibd}
\title{Make IBD (individual energy)}
\usage{
make_ibd(an_isd)
}
\arguments{
\item{an_isd}{an isd with column ind_size}
}
\value{
an_isd with added column for energy
}
\description{
Make IBD (individual energy)
}
