% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_isd.R
\name{simulate_isd}
\alias{simulate_isd}
\title{Simulate an ISD for a single year}
\usage{
simulate_isd(abundances, sd_dat = NULL, use_th_pars = F, isd_seed = NULL)
}
\arguments{
\item{abundances}{long dataframe of `id`, `abundance`, and any other identifying columns (such as `year`)}

\item{sd_dat}{optionally pass table with `id`, `mean_mass`, `mean_sd`}

\item{use_th_pars}{default FALSE, use Thibault (2010) parameter estiamtes?}

\item{isd_seed}{optionally specify seed to use}
}
\value{
dataframe with one row per simulated individual, columns `id`, `mass`, `isd_seed`, and any columns in `abundances`
}
\description{
Just using the abundances.
}
