% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_isd.R
\name{simulate_isd_route}
\alias{simulate_isd_route}
\title{Simulate the ISD for a BBS route}
\usage{
simulate_isd_route(
  route,
  region,
  sd_dat = NULL,
  censusyears = NULL,
  use_th_pars = F,
  isd_seed = NULL
)
}
\arguments{
\item{route}{which route}

\item{region}{which region}

\item{sd_dat}{optional, provide dataframe with `id`, `mean_mass`, and `mean_sd`}

\item{censusyears}{optional, provide a censusyear or years to use. If not provided returns for all years}

\item{use_th_pars}{default FALSE, whether to use sds estimated with Thibault (2010) parameters. Values are very close.}

\item{isd_seed}{optional, seed to use when simulating the ISD. If not provided, a seed is randomly chosen and returned with the simulated ISDs.}
}
\value{
list of `isd`, `covariates`, and `metadata`
}
\description{
Assumes MATSS is installed and working.
}
\details{
Uses `MATSS::get_bbs_route_region_data` to pull in a dataset and then simulates the ISD for a given year, or all years, for that dataset. Assumes MATSS is installed and working and `MATSS::prepare_bbs_data` has already run.
}
