% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_isd.R
\name{simulate_isd_ts}
\alias{simulate_isd_ts}
\title{Simulate ISD data for a timeseries}
\usage{
simulate_isd_ts(
  dataset,
  sd_dat = NULL,
  censusyears = NULL,
  use_th_pars = F,
  isd_seed = NULL
)
}
\arguments{
\item{dataset}{list of `abundance`, `covariates`, `metadata`. See `MATSS`.}

\item{sd_dat}{optional, provide a dataframe with `id`, `mean_mass`, `mean_sd`}

\item{censusyears}{optional, specify the year to simulate for. If not provided simulates for all years.}

\item{use_th_pars}{default F, use Thibault (2010) scaling parameters or re-estimated ones?}

\item{isd_seed}{optionally specify the seed to use when simulating the ISDs. If not provided one will be chosen and returned with output.}
}
\value{
list of `isd`, `covariates`, `metadata`
}
\description{
For one year or all years.
}
