% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging_methods.R
\name{summarize_over_individuals}
\alias{summarize_over_individuals}
\title{Summarize over individuals}
\usage{
summarize_over_individuals(an_ibd, currency)
}
\arguments{
\item{an_ibd}{cols for ind_size, ind_b}

\item{currency}{"mass" or "energy"}
}
\value{
species totals
}
\description{
To create a MATSS compatible dataset
}
