% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging_methods.R
\name{swap_currencies}
\alias{swap_currencies}
\title{Swap currencies}
\usage{
swap_currencies(a_dataset, an_ibd, currency)
}
\arguments{
\item{a_dataset}{a MATSS dataset}

\item{an_ibd}{an ibd}

\item{currency}{which currency to put in}
}
\value{
MATSS dataset with different currency
}
\description{
Swap currencies
}
